fix(unity-bootstrap-theme): move ranking cards into card group
