+++ packages/unity-bootstrap-theme/stories/molecules/cards/ranking-cards.js
@@ -4,6 +4,7 @@ import { googleAnalytics } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 export const rankingFunc = () => {
   const $infoLayer = document.querySelector(".info-layer");
   const $toggleIcon = document.getElementById("dispatch");
+alert('hey');
 
   $toggleIcon.addEventListener("click", function () {
     $infoLayer.classList.toggle("active");
